refactor(dispatchers): move scheduler/dispatcher to csrc/dispatchers, rename MPIScheduler
Declutter neml2/csrc/aoti/ -- the dispatcher subsystem (WorkScheduler,
SimpleScheduler, MPISimpleScheduler, DispatchedModel, batch_chunk, the
load_model factory) moves to a sibling neml2/csrc/dispatchers/ folder. Still
one library (libneml2_aoti) and one namespace (neml2::aoti); aoti/ keeps only
the AOTI Model runtime (Model, ops/solve/jacobian, Newton, the pybind binding).

Also rename MPIScheduler -> MPISimpleScheduler (it is the one-GPU-per-rank
*simple* scheduler; clearer alongside the upcoming async StaticHybridScheduler).

Both are pre-release organizational changes: paths/headers/CMake sources +
shipped FILE_SET, tests/cpp includes, and the dispatcher/migration docs are
updated to match. No behavior change -- the full dispatcher ctest suite and the
Release wheel build pass.
