+++ doc/content/migration/212_300.md
@@ -191,7 +191,7 @@ from your input files — the v2 HIT surface is gone. In v3 dispatch is
 a feature of the compiled C++ runtime: `neml2-compile --device cpu cuda`
 emits one artifact per device, and `neml2::aoti::load_model(stub, name,
 scheduler)` runs a batch across them via `SimpleScheduler` /
-`MPIScheduler`, configured in C++ source rather than the `.i`. See
+`MPISimpleScheduler`, configured in C++ source rather than the `.i`. See
 [](model-dispatch).
 
 Python (`neml2.load_model`, `neml2-run`) stays eager and single-device;
