+++ neml2/csrc/dispatchers/factory.cpp
@@ -28,9 +28,9 @@
 #include <nmhit/nmhit.h>
 
 #include "neml2/csrc/aoti/Model.h"
-#include "neml2/csrc/aoti/SimpleScheduler.h"
+#include "neml2/csrc/dispatchers/SimpleScheduler.h"
 #include "neml2/csrc/aoti/assertions.h"
-#include "neml2/csrc/aoti/factory.h"
+#include "neml2/csrc/dispatchers/factory.h"
 
 namespace neml2::aoti
 {
